Make pages and words follow the vehicle kind; intercompany from the fund's side
- Navigation: an SPV loses Portfolio construction; a GP entity loses the
  schedule of investments, construction and the fund-of-funds pages
  (lib/accounting/nav.ts hideFor, on the mechanism phase 5 added).
- Vocabulary: the statements page, the workbook and the PDF take the equity
  label from the payload (partners' capital, members' capital, owner's
  equity) and leave out the changes-in-capital statement where there are no
  partners. The setup checklist says what the close does for the kind.
  Admin's vehicle card uses VEHICLE_KIND_LABELS instead of its own copy.
- Intercompany: the panel moves to components/accounting/intercompany-panel
  (the manco page re-exports it) and takes a perspective — the same rows in
  a fund's words: they owe us / we owe them, they billed us / we billed
  them — with the counterparty fixed to the fund. A fund's or GP entity's
  Admin page shows one panel per management company, for callers holding
  the management-company grant, via app/(app)/funds/status/intercompany-card.
- Not done, on purpose: an SPV chart preset. The book-to-tax run refuses to
  post without the organizational, syndication and deferred-cost accounts,
  so trimming them would break an SPV's tax year. Recorded in CHANGELOG.
