+++ CHANGELOG.md
@@ -19,6 +19,9 @@ Added
 - **The `individual` kind** (migration `20260904000004`): an angel, or their single-member LLC. Its own chart — a fund's investment side code for code, owner's capital and draws, and a personal-return expense chart — hidden partner pages (capital accounts, portfolio construction, fund-of-funds), and the balance sheet's equity labelled Owner's equity (Members' capital on a manco; `lib/accounting/vocab.ts`)
 - **Realized gains by lot** (`/api/accounting/realized-gains`, on the Tax page, and `realized-gains-<year>.csv` in the tax package): each disposal, the lots it consumed under the fund's lot method, proceeds and gain apportioned per lot, short- or long-term
 - The opening-balance bootstrap seeds the chart for the vehicle's kind instead of always the fund chart
+- **Kind-aware pages and words.** An SPV no longer shows Portfolio construction; a GP entity no longer shows the schedule of investments, construction or the fund-of-funds pages; the balance sheet, the workbook, the PDF and the statements page name the equity for what it is (partners' capital, members' capital, owner's equity) and leave out the changes-in-capital statement where there are no partners; the setup checklist says what the close does for the kind; Admin's vehicle card uses the one kind vocabulary instead of its own copy
+- **Intercompany from the fund's side.** The intercompany panel is shared (`components/accounting/intercompany-panel.tsx`) and a fund's or GP entity's Admin page shows its charges with the management company from its own point of view — "we owe them / they owe us", "they billed us / we billed them" — with the same record and settle actions. Needs the management-company grant, as the routes do; the card renders nothing without it or without a management company
+- Not done: an SPV chart preset. The book-to-tax run requires the organizational, syndication and deferred-cost accounts to exist before it will post, so trimming them from an SPV's chart would break its tax year; the SPV keeps the fund chart
 - Nav sections can name the kinds they are hidden for; the vehicle context now carries the kind
 - **Tax page** at Funds → Tax (behind the tax reporting feature): year status with close and reopen, the derived adjustments with Post to the tax book, the adjusting entries in the books, K-1 package generate / finalise / amend with the workbook download and state worklist, partner tax forms, K-1s owed by underlying funds, and the tax package. The routes existed; the page did not
 - **The workbook's GL Detail now rolls each account** — opening balance carried in at the period start, each posting with the accounts on the other side, running balance, closing balance equal to the trial balance — instead of listing period activity with no balances. The Trial Balance and Schedule of Investments tabs gain a column pair per compared period, like the statements already had
